Refactor ItemTaskViewHolder and ItemBoardListViewHolder drag listeners.
Implement FirestoreRepository methods for updating tasks positions based
on the requierement (move, delete, insert).
Extract drag related callbacks into separate interfaces.

+++ app/src/androidTest/java/com/example/kanbun/domain/FirestoreTestUtil.kt
@@ -128,7 +128,7 @@ class FirestoreTestUtil {
         fun createBoardList(name: String, position: Int): BoardList =
             BoardList(
                 name = name,
-                position = position,
+                position = position.toLong(),
                 tasks = emptyList()
             )
